Rename back to procedure mapper
