+++ procedural/constraints/src/main/kotlin/gov/nasa/ammos/aerie/procedural/constraints/ProcedureMapper.kt
@@ -1,9 +1,9 @@
-package gov.nasa.ammos.aerie.procedural.scheduling
+package gov.nasa.ammos.aerie.procedural.constraints
 
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue
 import gov.nasa.jpl.aerie.merlin.protocol.types.ValueSchema
 
-interface SchedulingProcedureMapper<T: Goal> {
+interface ProcedureMapper<T: Constraint> {
   fun valueSchema(): ValueSchema
   fun serialize(procedure: T): SerializedValue
   fun deserialize(arguments: SerializedValue): T
