+++ procedural/scheduling/src/main/kotlin/gov/nasa/ammos/aerie/procedural/scheduling/ProcedureMapper.kt
@@ -1,9 +1,9 @@
-package gov.nasa.ammos.aerie.procedural.constraints
+package gov.nasa.ammos.aerie.procedural.scheduling
 
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue
 import gov.nasa.jpl.aerie.merlin.protocol.types.ValueSchema
 
-interface ConstraintProcedureMapper<T: Constraint> {
+interface ProcedureMapper<T: Goal> {
   fun valueSchema(): ValueSchema
   fun serialize(procedure: T): SerializedValue
   fun deserialize(arguments: SerializedValue): T
